pyproject.toml
[tool.poetry]
name = "aioworkers-loki"
version = "0.0.0"
description = ""
authors = ["Alexander Malev <malev@somedev.ru>"]
license = "Apache Software License 2.0"
readme = "README.rst"
homepage = "https://github.com/aioworkers/aioworkers-loki"
repository = "https://github.com/aioworkers/aioworkers-loki"

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Framework :: AsyncIO",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]

packages = [
    { include = "aioworkers_loki" },
]

[tool.poetry.dependencies]
python = "^3.6"
aioworkers = "^0.18"
python-logging-loki = "^0.3.1"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
flake8 = "^3.9.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = true
# The pattern for git tag in pep440 format. It is the same as default but without the v prefix.
pattern = "(?x)^(?P<base>\\d+\\.\\d+\\.\\d+)(-?((?P<stage>[a-zA-Z]+)\\.?(?P<revision>\\d+)?))?(\\+(?P<tagged_metadata>.+))?$"

[tool.poetry-dynamic-versioning.substitution]
enable = true
